Signin responsive design
