+++ packages/messenger-widget/src/components/SignIn/SignIn.css
@@ -131,3 +131,5 @@
         padding: 0rem 8rem 0rem 8rem !important;
     }
 }
+
+/* ==================== Mobile Responsive CSS ==================== */
